requirements.txt
click~=6.0
dnspython>=1.13.0
ldap3~=2.2
pyasn1~=0.2.0

pytest
pytest-cov>=2.0.0
pytest-runner>=2.0.0
